Minor renaming of arrival page
