+++ app/src/main/java/com/lmorda/shopper/arrival/ArrivalViewModel.kt
@@ -1,11 +1,11 @@
-package com.lmorda.shopper.order
+package com.lmorda.shopper.arrival
 
 import androidx.lifecycle.*
 import com.lmorda.shopper.ORDER_ID_ARG
 import com.lmorda.shopper.data.ShopperRepository
 
-class OrderDetailsViewModel(val shopperRepository: ShopperRepository,
-                            val state: SavedStateHandle): ViewModel() {
+class ArrivalViewModel(val shopperRepository: ShopperRepository,
+                       val state: SavedStateHandle): ViewModel() {
 
     var orderId: Int? = -1
