+++ app/src/main/res/layout/fragment_arrival.xml
