+++ app/src/main/res/navigation/nav_graph.xml
@@ -81,9 +81,9 @@
 
     <fragment
         android:id="@+id/orderFragment"
-        android:name="com.lmorda.shopper.order.OrderDetailsFragment"
+        android:name="com.lmorda.shopper.arrival.ArrivalFragment"
         android:label="OrderFragment"
-        tools:layout="@layout/fragment_order">
+        tools:layout="@layout/fragment_arrival">
 
         <action
             android:id="@+id/action_orderFragment_to_storeFragment"
